Overhaul wxMesh to specifically pair with weewx-mqtt extension.
This driver is made to work as a receiver of data published from the weewx-mqtt extension only. 

This is because weewx-mqtt publishes data with `observation_unitLabel`. So this driver is looking for that, trimming the `_unitLabel` off the key name, then sends the key and value to the loop untouched. 

Because of this, this driver makes a great candidate to get a lab system going that has mirrored “production” data. 

* Production weewx will publish data to MQTT using the `weewx-mqtt` extension. 
* The lab system will use this driver to subscribe to that data and process it as LOOP data.